competition/202206.tanachou/sample_submit_paddle/requirements.txt
astor==0.8.1
Babel==2.9.1
bce-python-sdk==0.8.64
cachetools==5.0.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.2
cssselect==1.1.0
cssutils==2.4.0
cycler==0.11.0
Cython==0.29.28
decorator==5.1.1
distlib==0.3.4
et-xmlfile==1.1.0
filelock==3.6.0
flake8==4.0.1
Flask==2.1.1
Flask-Babel==2.0.0
fonttools==4.32.0
future==0.18.2
identify==2.4.12
idna==3.3
imageio==2.17.0
imgaug==0.4.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
lmdb==1.3.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.6.1
networkx==2.8
nodeenv==1.6.0
numpy==1.22.3
opencv-contrib-python==4.4.0.46
openpyxl==3.0.9
packaging==21.3
# paddle
--find-links  https://www.paddlepaddle.org.cn/whl/linux/mkl/avx/stable.html
paddlepaddle-gpu==2.2.2.post112
paddleocr==2.4.0.4
pandas==1.4.2
Pillow==9.1.0
platformdirs==2.5.2
pre-commit==2.18.1
premailer==3.10.0
protobuf==3.20.0
pyclipper==1.3.0.post2
pycodestyle==2.8.0
pycryptodome==3.14.1
pyflakes==2.4.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
Shapely==1.8.1.post1
shellcheck-py==0.8.0.4
six==1.16.0
tifffile==2022.4.8
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.9
virtualenv==20.14.1
visualdl==2.2.3
Werkzeug==2.1.1
zipp==3.8.0
prettytable
ujson
opencv-python==4.4.0.46
scipy
scikit-learn>=0.21.0
gast==0.3.3
faiss-cpu==1.7.1.post2
easydict
